requirements.txt
asgiref==3.2.5
Django==3.0.4
djangorestframework==3.11.0
pytz==2019.3
selenium==3.141.0
sqlparse==0.3.1
urllib3==1.25.8
